Изменить структуру проекта
