--- src/Enums/TaskStatus.java
@@ -1,7 +1,0 @@
-package Enums;
-
-public enum TaskStatus {
-    NEW,
-    IN_PROGRESS,
-    DONE
-}
