--- src/Enums/TaskType.java
@@ -1,7 +1,0 @@
-package Enums;
-
-public enum TaskType {
-    TASK,
-    EPIC,
-    SUBTASK
-}
