--- src/Managers/HistoryManager.java
@@ -1,12 +1,0 @@
-package Managers;
-
-import java.util.List;
-import Tasks.Task;
-
-public interface HistoryManager {
-    void add(Task task);
-
-    void remove(int id);
-
-    List<Task> getHistory();
-}
