--- src/Managers/Managers.java
@@ -1,11 +1,0 @@
-package Managers;
-
-public class Managers {
-    public static TaskManager getDefault() {
-        return new InMemoryTaskManager();
-    }
-
-    public static HistoryManager getDefaultHistory() {
-        return new InMemoryHistoryManager();
-    }
-}
