--- src/Node/Node.java
@@ -1,16 +1,0 @@
-package Node;
-
-public class Node<T> {
-    public T data;
-    public Node<T> next;
-    public Node<T> prev;
-
-    public Node() {
-    }
-
-    public Node(T data) {
-        this.data = data;
-        this.next = null;
-        this.prev = null;
-    }
-}
