--- src/SaveExceptions/ManagerSaveException.java
@@ -1,11 +1,0 @@
-package SaveExceptions;
-
-public class ManagerSaveException extends RuntimeException {
-    public ManagerSaveException() {
-
-    }
-
-    public ManagerSaveException(String par) {
-        System.out.println(par);
-    }
-}
