+++ src/managers/inmemory/Managers.java
@@ -1,0 +1,14 @@
+package managers.inmemory;
+
+import managers.interfaces.HistoryManager;
+import managers.interfaces.TaskManager;
+
+public class Managers {
+    public static TaskManager getDefault() {
+        return new InMemoryTaskManager();
+    }
+
+    public static HistoryManager getDefaultHistory() {
+        return new InMemoryHistoryManager();
+    }
+}
