+++ src/save/exception/ManagerSaveException.java
@@ -1,0 +1,11 @@
+package save.exception;
+
+public class ManagerSaveException extends RuntimeException {
+    public ManagerSaveException() {
+
+    }
+
+    public ManagerSaveException(String par) {
+        System.out.println(par);
+    }
+}
