+++ src/task/enums/TaskStatus.java
@@ -1,0 +1,7 @@
+package task.enums;
+
+public enum TaskStatus {
+    NEW,
+    IN_PROGRESS,
+    DONE
+}
