+++ src/task/enums/TaskType.java
@@ -1,0 +1,7 @@
+package task.enums;
+
+public enum TaskType {
+    TASK,
+    EPIC,
+    SUBTASK
+}
