+++ src/task/node/Node.java
@@ -1,0 +1,16 @@
+package task.node;
+
+public class Node<T> {
+    public T data;
+    public Node<T> next;
+    public Node<T> prev;
+
+    public Node() {
+    }
+
+    public Node(T data) {
+        this.data = data;
+        this.next = null;
+        this.prev = null;
+    }
+}
