+++ src/tasks/epic/Epic.java
@@ -1,0 +1,50 @@
+package tasks.epic;
+
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import task.enums.TaskStatus;
+import tasks.subtask.SubTask;
+import tasks.task.Task;
+
+public class Epic extends Task {
+    private final List<SubTask> subTasks;
+    private LocalDateTime endTime;
+
+    public Epic(String newEpicName, String newDescription) {
+        super(newEpicName, newDescription, TaskStatus.NEW, Duration.ZERO, LocalDateTime.MIN);
+        endTime = LocalDateTime.MIN;
+        subTasks = new ArrayList<>();
+    }
+
+    public void addSubTaskToEpic(SubTask newSubTask) {
+        subTasks.add(newSubTask);
+
+        if (subTasks.size() < 2) {
+            this.setStartTime(newSubTask.getStartTime());
+            this.setDuration(newSubTask.getDuration());
+            endTime = newSubTask.getEndTime();
+            return;
+        }
+
+        for (SubTask task : subTasks) {
+            this.setDuration(this.getDuration().plus(task.getDuration()));
+
+            if (task.getStartTime().isBefore(this.getStartTime())) {
+                this.setStartTime(task.getStartTime());
+            } else if (task.getEndTime().isAfter(endTime)) {
+                endTime = task.getEndTime();
+            }
+        }
+    }
+
+    public List<SubTask> getSubTasks() {
+        return subTasks;
+    }
+
+    @Override
+    public LocalDateTime getEndTime() {
+        return endTime;
+    }
+}
