+++ test/HttpTaskServerTest.java
@@ -1,11 +1,11 @@
-import CustomFormatAdapters.FormatAdapters;
-import HttpServer.HttpTaskServer;
-import Managers.InMemoryTaskManager;
-import Managers.TaskManager;
-import Tasks.Epic;
-import Tasks.SubTask;
-import Tasks.Task;
-import Enums.TaskStatus;
+import custom.format.adapters.FormatAdapters;
+import http.server.HttpTaskServer;
+import managers.inmemory.InMemoryTaskManager;
+import managers.interfaces.TaskManager;
+import tasks.epic.Epic;
+import tasks.subtask.SubTask;
+import tasks.task.Task;
+import task.enums.TaskStatus;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonIOException;
