+++ test/TaskManagerTest.java
@@ -1,8 +1,8 @@
-import Managers.TaskManager;
-import Tasks.Epic;
-import Tasks.SubTask;
-import Tasks.Task;
-import Enums.TaskStatus;
+import managers.interfaces.TaskManager;
+import tasks.epic.Epic;
+import tasks.subtask.SubTask;
+import tasks.task.Task;
+import task.enums.TaskStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
